+++ spec/models/audit/event/check_spec.rb
@@ -3,31 +3,33 @@ require 'spec_helper'
 describe Audit::Event::Check do
 
   let(:user) { double('my-user', id: 'rspec:user:my_user') }
-  let(:resource) { double('my-resource', id: 'rspec:variable:my_var') }
+  let(:resource_id)  {'rspec:variable:my_var'}
   let(:privilege) { 'execute' }
-  let(:role) { double('my-host', id: 'rspec:host:my_host') }
+  let(:role_id) { 'rspec:host:my_host' }
   let(:client_ip) { 'my-client-ip' }
   let(:success) { true }
   let(:operation) { 'check' }
+  let(:error_message) { nil }
 
   subject do
     Audit::Event::Check.new(
       user: user,
-      resource: resource,
+      resource_id: resource_id,
       privilege: privilege,
-      role: role,
+      role_id: role_id,
       client_ip: client_ip,
       operation: operation,
-      success: success
+      success: success,
+      error_message: error_message
     )
   end
 
   context 'when successful' do
     it 'produces the expected message' do
       expect(subject.message).to eq(
-        'rspec:user:my_user checked if rspec:host:my_host can execute ' \
-        'rspec:variable:my_var (success)'
-      )
+        'rspec:user:my_user successfully checked if rspec:host:my_host can execute ' \
+        'rspec:variable:my_var '
+                                 )
     end
 
     it 'uses the INFO log level' do
@@ -36,9 +38,9 @@ describe Audit::Event::Check do
 
     it 'renders to string correctly' do
       expect(subject.to_s).to eq(
-        'rspec:user:my_user checked if rspec:host:my_host can execute ' \
-        'rspec:variable:my_var (success)'
-      )
+        'rspec:user:my_user successfully checked if rspec:host:my_host can execute ' \
+        'rspec:variable:my_var '
+                              )
     end
 
     it 'produces the expected action_sd' do
@@ -53,9 +55,9 @@ describe Audit::Event::Check do
 
     it 'produces the expected message' do
       expect(subject.message).to eq(
-        'rspec:user:my_user checked if rspec:host:my_host can execute ' \
-        'rspec:variable:my_var (failure)'
-      )
+        'rspec:user:my_user failed to check if rspec:host:my_host can execute ' \
+        'rspec:variable:my_var '
+                                 )
     end
 
     it 'uses the WARNING log level' do
@@ -71,12 +73,13 @@ describe Audit::Event::Check do
 
   context 'when the resource does not exist and a failure occurs' do
     let(:success) { false }
-    let(:resource) { 'rspec:variable:non_existing_var' }
+    let(:resource_id) { 'rspec:variable:non_existing_var' }
+    let(:error_message) { 'Variable \'non-existing-var\' not found in account \'CyberArk\''}
 
     it 'produces the expected message' do
       expect(subject.message).to eq(
-                                   'rspec:user:my_user checked if rspec:host:my_host can execute ' \
-        'rspec:variable:non_existing_var (failure)'
+        'rspec:user:my_user failed to check if rspec:host:my_host can execute ' \
+        'rspec:variable:non_existing_var : Variable \'non-existing-var\' not found in account \'CyberArk\''
                                  )
     end
 
@@ -85,12 +88,13 @@ describe Audit::Event::Check do
 
   context 'when the role does not exist and a failure occurs' do
     let(:success) { false }
-    let(:role) { 'rspec:host:my_non_existing_host' }
+    let(:role_id) { 'rspec:host:my_non_existing_host' }
+    let(:error_message) { 'Forbidden' }
 
     it 'produces the expected message' do
       expect(subject.message).to eq(
-                                   'rspec:user:my_user checked if rspec:host:my_non_existing_host can execute ' \
-        'rspec:variable:my_var (failure)'
+        'rspec:user:my_user failed to check if rspec:host:my_non_existing_host can execute ' \
+        'rspec:variable:my_var : Forbidden'
                                  )
     end